Fixed the geolocation error + some ease of life updates
Yes this is a late commit, yes I have checked that it works mutliple times. Note:s 2 Noticable Bugs remain:
- You cannot got a geolcation search and then do an address search as I haven't made a way of toggling this yet.

- The table takes some time to load all the results from the query into the table which I discovered was not working properly. (This was what was actually causing the geolocation error in the end as it somehow kept it's own personal global scope util the very end of the forEach loop and so never updated the userGeolocation variable)

As I was doing this to keep my sanity I updated some of the things that have been bugging me such as the removable of the zip code and state fields in the table and the proper resetting of the search fields.